+++ src/.gitignore
@@ -1,3 +1,3 @@
 *.class
-acpcommander
 *.jar
+out
